+++ src/cross-fetch.ts
@@ -1,0 +1,6 @@
+/**
+ * This file wraps cross-platform npm modules for 
+ * correct processing of vite-plugin-cross-platform
+ */
+import fetch from "cross-fetch";
+export default fetch;