+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## TODO
+
+- Enable usage of extension on vscode for web
+
+## Unreleased
+
+- Added Wikidot syncing
+
 ## 0.0.1: Pre-Alpha 1
 
 - Initial release